move rest out of peripherals
