+++ packages/backend/src/config/Config.ts
@@ -12,13 +12,13 @@ import type {
 } from '@l2beat/discovery'
 import type { TrackedTxConfigEntry } from '@l2beat/shared'
 import type { CoingeckoId, ProjectId, UnixTime } from '@l2beat/shared-pure'
+import type { MulticallConfigEntry } from '../modules/tvs/tools/sharedEscrows/multicall/types'
 import type {
   AmountConfig,
   BlockTimestampConfig,
   PriceConfig,
   ProjectTvsConfig,
 } from '../modules/tvs/types'
-import type { MulticallConfigEntry } from '../peripherals/multicall/types'
 import type { ChainApi } from './chain/ChainApi'
 import type { ResolvedFeatureFlag } from './FeatureFlags'
