+++ packages/backend/src/modules/tvs/tools/sharedEscrows/getAggLayerTokens.ts
@@ -14,14 +14,14 @@ import {
   TokenId,
 } from '@l2beat/shared-pure'
 import { utils } from 'ethers'
-import { MulticallClient } from '../../../../peripherals/multicall/MulticallClient'
-import { toMulticallConfigEntry } from '../../../../peripherals/multicall/MulticallConfig'
-import type { MulticallRequest } from '../../../../peripherals/multicall/types'
 import { getTimeRangeIntersection } from '../getTimeRangeIntersection'
 import type { LocalStorage } from '../LocalStorage'
 import { createEscrowToken } from '../legacyConfig/mapLegacyConfig'
 import type { LegacyEscrow } from '../legacyConfig/types'
 import { isEmptyAddress } from './isEmptyAddress'
+import { MulticallClient } from './multicall/MulticallClient'
+import { toMulticallConfigEntry } from './multicall/MulticallConfig'
+import type { MulticallRequest } from './multicall/types'
 
 export const bridgeInterface = new utils.Interface([
   'function getTokenWrappedAddress(uint32 originNetwork, address originTokenAddress) view returns (address)',
